fix: add state_provier as a class in flutter_searchbox lib
